# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lab Resource Utilization Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/profile:
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── INSTITUTIONS ──────────────────────────────────────────────────────────
...
  /analytics/equipment-utilization:
    get:
      operationId: getEquipmentUtilizationReport
      tags: [analytics]
      summary: Get equipment utilization analytics
      parameters:
        - name: period
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Utilization analytics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentUtilizationReport"
  /analytics/booking-trends:
    get:
      operationId: getBookingTrends
      tags: [analytics]
      summary: Get booking trends over time
      parameters:
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Booking trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingTrendPoint"
  /analytics/top-equipment:
    get:
      operationId: getTopEquipment
      tags: [analytics]
      summary: Get most-used equipment
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopEquipmentItem"
  /analytics/cost-analysis:
    get:
      operationId: getCostAnalysis
      tags: [analytics]
      summary: Get cost analysis by department
      parameters:
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Cost analysis data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CostAnalysisItem"
...
  /users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: institutionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [auth]
      summary: Update user role
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [RESEARCHER, LAB_TECHNICIAN, LAB_MANAGER, DEPT_HEAD, INSTITUTION_ADMIN, SYSTEM_ADMIN]
        institutionId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - firstName
        - lastName
        - role
        - createdAt
    ProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    UserRoleUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [RESEARCHER, LAB_TECHNICIAN, LAB_MANAGER, DEPT_HEAD, INSTITUTION_ADMIN, SYSTEM_ADMIN]
        institutionId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
      required:
        - role
...
    EquipmentUtilizationReport:
      type: object
      properties:
        equipmentId:
          type: integer
        equipmentName:
          type: string
        departmentName:
          type: string
        utilizationRate:
          type: number
        totalHoursUsed:
          type: number
        bookingCount:
          type: integer
        revenue:
          type: number
      required:
        - equipmentId
        - equipmentName
        - departmentName
        - utilizationRate
        - totalHoursUsed
        - bookingCount
        - revenue
    BookingTrendPoint:
      type: object
      properties:
        date:
          type: string
        bookingCount:
          type: integer
        confirmedCount:
          type: integer
        cancelledCount:
          type: integer
      required:
        - date
        - bookingCount
        - confirmedCount
        - cancelledCount
    TopEquipmentItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        model:
          type: string
        departmentName:
          type: string
        bookingCount:
          type: integer
        utilizationRate:
          type: number
        revenue:
          type: number
      required:
        - id
        - name
        - model
        - departmentName
        - bookingCount
        - utilizationRate
        - revenue
    CostAnalysisItem:
      type: object
      properties:
        departmentId:
          type: integer
        departmentName:
          type: string
        totalCost:
          type: number
        equipmentCount:
          type: integer
        bookingCount:
          type: integer
      required:
        - departmentId
        - departmentName
        - totalCost
        - equipmentCount
        - bookingCount
...