# lib/api-zod/src/generated/types/bookingTrendPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Lab Resource Utilization Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingTrendPoint {
  date: string;
  bookingCount: number;
  confirmedCount: number;
  cancelledCount: number;
}